requirements.txt
# Core dependencies for Railway deployment
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20
python-dotenv==1.2.1
starlette==0.37.2
pydantic==2.12.3
requests==2.32.5

# AI - Gemini API
google-generativeai==0.8.3

# CAD Engine - Full functionality
cadquery==2.6.1
ezdxf==1.4.3
